We weren't checking against the Person policy when displaying the alternate phone on search results. So alternate phone numbers were showing up on non-authenticated searches.
The logic was a little funky too, it was only showing alternative phones if the direct phone number was available, but that doesn't really make sense. Because it is possible for someone to only have an alternate phone number and no direct, in which case it wouldn't show up in the search results.